+++ src/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import {createStore} from "redux";
 import {Provider} from "react-redux";
-import {reducer} from "./store/reducer.js";
+import reducer from "./store/reducer.js";
 
 import App from "./components/app/app.jsx";
