+++ src/store/reducer.js
@@ -53,3 +53,19 @@ export const reducer = (state = initialState, action) => {
     default: return state;
   }
 };
+
+export const getStep = (state) => {
+  return state.step;
+};
+
+export const getMistakes = (state) => {
+  return state.mistakes;
+};
+
+export const getMaxMistakes = (state) => {
+  return state.maxMistakes;
+};
+
+export const getQuestions = (state) => {
+  return state.questions;
+};
