--- src/store/selectors.js
@@ -1,15 +1,0 @@
-export const getStep = (state) => {
-  return state.step;
-};
-
-export const getMistakes = (state) => {
-  return state.mistakes;
-};
-
-export const getMaxMistakes = (state) => {
-  return state.maxMistakes;
-};
-
-export const getQuestions = (state) => {
-  return state.questions;
-};
